Fixed 0xc0000005 error in berkelium_d.exe
This was caused by a missing reference to strcmp, and had something to do with linking to the wrong version of the CRT.

Note: To debug this, you need to set a breakpoint in BrokerServicesBase::SpawnTarget after target->Create. Then, attach to the newly spawned process, and break in ResolverThunk::ResolveInterceptor(), continue, and you will be able to debug the crash.