Use externalHost.postMessage instead of chrome.send
This solves the problem with POST, so you can now log into gmail.

Additionally, we now get passed an origin field. The target can be any arbitrary URL, including a urn, so it can be used as a sort of identifier.

There's no new Berkelium api yet, so the message uses onChromeSend with name "externalHost" and exactly 3 arguments.